refactor(cdk): reorganize logical IDs by resource groups
- Group CDK resources by logical function
- Improve resource discoverability in CloudFormation
- Make infrastructure components easier to locate and manage
- Establish consistent naming convention for resource groups
