+++ deploy/src/shared/index.ts
@@ -15,25 +15,18 @@
 
 import * as cdk from 'aws-cdk-lib';
 import { Construct } from 'constructs';
+import * as logs from 'aws-cdk-lib/aws-logs';
+import { RemovalPolicy } from 'aws-cdk-lib';
+import * as ec2 from 'aws-cdk-lib/aws-ec2';
+import { Vpc, ISubnet } from 'aws-cdk-lib/aws-ec2';
 
 /* eslint-disable @typescript-eslint/no-empty-interface */
 export interface CognitoWebNativeConstructProps extends cdk.StackProps {}
 
-const defaultProps: Partial<CognitoWebNativeConstructProps> = {};
-
-/**
- * Deploys Cognito with an Authenticated & UnAuthenticated Role with a Web and Native client
- */
-export class Authentication extends Construct {
-	public userPool: cdk.aws_cognito.UserPool;
-	public webClientUserPool: cdk.aws_cognito.UserPoolClient;
-	public nativeClientUserPool: cdk.aws_cognito.UserPoolClient;
-	public userPoolId: string;
-	public identityPoolId: string;
-	public webClientId: string;
-	public nativeClientId: string;
-	public authenticatedRole: cdk.aws_iam.Role;
-	public unauthenticatedRole: cdk.aws_iam.Role;
+export class Shared extends Construct {
+	public vpc: Vpc;
+	public subnets: ISubnet[];
+	public securityGroup: string;
 
 	constructor(
 		parent: Construct,
@@ -42,218 +35,54 @@ export class Authentication extends Construct {
 	) {
 		super(parent, name);
 
-		/* eslint-disable @typescript-eslint/no-unused-vars */
-		props = { ...defaultProps, ...props };
-
-		const userPool = new cdk.aws_cognito.UserPool(this, 'UserPool', {
-			signInAliases: {
-				email: true,
-				username: true,
-			},
-			selfSignUpEnabled: false, // required by Pallisade, if you enable it will create a sev 3/2 ticket
-			autoVerify: { email: true },
-			userVerification: {
-				emailSubject: 'Verify your email the app!',
-				emailBody:
-					'Hello {username}, Thanks for signing up to the app! Your verification code is {####}',
-				emailStyle: cdk.aws_cognito.VerificationEmailStyle.CODE,
-				smsMessage:
-					'Hello {username}, Thanks for signing up to app! Your verification code is {####}',
-			},
-			passwordPolicy: {
-				minLength: 8,
-				requireDigits: true,
-				requireUppercase: true,
-				requireSymbols: true,
-				requireLowercase: true,
-			},
-			removalPolicy: cdk.RemovalPolicy.DESTROY,
+		const vpcLogs = new logs.LogGroup(this, 'VpcFlowLogs', {
+			logGroupName: '/aws/vpc/flow-logs',
+			removalPolicy: RemovalPolicy.DESTROY,
 		});
 
-		const userPoolWebClient = new cdk.aws_cognito.UserPoolClient(
-			this,
-			'UserPoolWebClient',
-			{
-				generateSecret: false,
-				userPool: userPool,
-				userPoolClientName: 'WebClient',
-			}
-		);
+		const availabilityZones = cdk.Stack.of(this).availabilityZones;
 
-		const userPoolNativeClient = new cdk.aws_cognito.UserPoolClient(
-			this,
-			'UserPoolNativeClient',
-			{
-				generateSecret: true,
-				userPool: userPool,
-				userPoolClientName: 'NativeClient',
-			}
-		);
-
-		const identityPool = new cdk.aws_cognito.CfnIdentityPool(
-			this,
-			'IdentityPool',
-			{
-				allowUnauthenticatedIdentities: false,
-				cognitoIdentityProviders: [
-					{
-						clientId: userPoolWebClient.userPoolClientId,
-						providerName: userPool.userPoolProviderName,
-					},
-					{
-						clientId: userPoolNativeClient.userPoolClientId,
-						providerName: userPool.userPoolProviderName,
-					},
-				],
-			}
-		);
-
-		const unauthenticatedRole = new cdk.aws_iam.Role(
-			this,
-			'DefaultUnauthenticatedRole',
-			{
-				assumedBy: new cdk.aws_iam.FederatedPrincipal(
-					'cognito-identity.amazonaws.com',
-					{
-						StringEquals: {
-							'cognito-identity.amazonaws.com:aud':
-								identityPool.ref,
-						},
-						'ForAnyValue:StringLike': {
-							'cognito-identity.amazonaws.com:amr':
-								'unauthenticated',
-						},
-					},
-					'sts:AssumeRoleWithWebIdentity'
-				),
-			}
-		);
-
-		const authenticatedRole = new cdk.aws_iam.Role(
-			this,
-			'DefaultAuthenticatedRole',
-			{
-				assumedBy: new cdk.aws_iam.FederatedPrincipal(
-					'cognito-identity.amazonaws.com',
-					{
-						StringEquals: {
-							'cognito-identity.amazonaws.com:aud':
-								identityPool.ref,
-						},
-						'ForAnyValue:StringLike': {
-							'cognito-identity.amazonaws.com:amr':
-								'authenticated',
-						},
-					},
-					'sts:AssumeRoleWithWebIdentity'
-				),
-			}
-		);
-
-		new cdk.aws_cognito.CfnIdentityPoolRoleAttachment(
-			this,
-			'IdentityPoolRoleAttachment',
-			{
-				identityPoolId: identityPool.ref,
-				roles: {
-					unauthenticated: unauthenticatedRole.roleArn,
-					authenticated: authenticatedRole.roleArn,
+		const vpc = new ec2.Vpc(this, `Vpc`, {
+			ipAddresses: ec2.IpAddresses.cidr('10.0.0.0/16'),
+			natGateways: 1,
+			maxAzs: availabilityZones.length,
+			flowLogs: {
+				s3: {
+					destination:
+						ec2.FlowLogDestination.toCloudWatchLogs(vpcLogs),
+					trafficType: ec2.FlowLogTrafficType.ALL,
 				},
-			}
-		);
-
-		const adminGroup = new cdk.aws_cognito.CfnUserPoolGroup(
-			this,
-			'SdasAdminsGroup',
-			{
-				groupName: 'sdas-admins',
-				userPoolId: userPool.userPoolId,
-				description: 'SDAS Administrators',
-			}
-		);
-
-		// Admin User
-		const adminUser = new cdk.aws_cognito.CfnUserPoolUser(
-			this,
-			'SdasAdminUser',
-			{
-				userPoolId: userPool.userPoolId,
-				username: 'admin',
-				userAttributes: [
-					{
-						name: 'email',
-						value: this.node.tryGetContext('admin_email'),
-					},
-					{
-						name: 'given_name',
-						value: 'SDAS',
-					},
-					{
-						name: 'family_name',
-						value: 'Admin',
-					},
-					{
-						name: 'email_verified',
-						value: 'true',
-					},
-				],
-				desiredDeliveryMediums: ['EMAIL'],
-			}
-		);
-
-		const adminGroupAttachment =
-			new cdk.aws_cognito.CfnUserPoolUserToGroupAttachment(
-				this,
-				'AdminGroupAttachment',
+			},
+			subnetConfiguration: [
 				{
-					username: adminUser.username!,
-					groupName: adminGroup.groupName!,
-					userPoolId: userPool.userPoolId,
-				}
-			);
-
-		adminGroupAttachment.node.addDependency(adminGroup, adminUser);
-
-		// Assign Cfn Outputs
-		new cdk.CfnOutput(this, 'UserPoolId', {
-			value: userPool.userPoolId,
-		});
-		new cdk.CfnOutput(this, 'IdentityPoolId', {
-			value: identityPool.ref,
-		});
-		new cdk.CfnOutput(this, 'WebClientId', {
-			value: userPoolWebClient.userPoolClientId,
-		});
-		new cdk.CfnOutput(this, 'NativeClientId', {
-			value: userPoolNativeClient.userPoolClientId,
+					name: 'private-subnet-',
+					subnetType: ec2.SubnetType.PRIVATE_WITH_EGRESS,
+					cidrMask: 24,
+				},
+				{
+					name: 'public-subnet-',
+					subnetType: ec2.SubnetType.PUBLIC,
+					cidrMask: 24,
+				},
+			],
 		});
 
-		// Add SSM Parameters
-		new cdk.aws_ssm.StringParameter(this, 'COGNITO_USER_POOL_ID', {
-			stringValue: userPool.userPoolId,
+		new ec2.GatewayVpcEndpoint(this, 'VpcEndpointS3', {
+			service: ec2.GatewayVpcEndpointAwsService.S3,
+			vpc,
 		});
 
-		new cdk.aws_ssm.StringParameter(this, 'COGNITO_IDENTITY_POOL_ID', {
-			stringValue: identityPool.ref,
-		});
+		const subnets = vpc.privateSubnets;
+		const securityGroup = vpc.vpcDefaultSecurityGroup;
 
-		new cdk.aws_ssm.StringParameter(this, 'COGNITO_WEB_CLIENT_ID', {
-			stringValue: userPoolWebClient.userPoolClientId,
-		});
-
-		new cdk.aws_ssm.StringParameter(this, 'COGNITO_NATIVE_CLIENT_ID', {
-			stringValue: userPoolNativeClient.userPoolClientId,
+		new cdk.CfnOutput(this, 'VpcId', {
+			value: vpc.vpcId,
+			exportName: 'VpcId',
 		});
 
 		// assign public properties
-		this.userPool = userPool;
-		this.webClientUserPool = userPoolWebClient;
-		this.nativeClientUserPool = userPoolNativeClient;
-		this.authenticatedRole = authenticatedRole;
-		this.unauthenticatedRole = unauthenticatedRole;
-		this.userPoolId = userPool.userPoolId;
-		this.identityPoolId = identityPool.ref;
-		this.webClientId = userPoolWebClient.userPoolClientId;
-		this.nativeClientId = userPoolNativeClient.userPoolClientId;
+		this.vpc = vpc;
+		this.subnets = subnets;
+		this.securityGroup = securityGroup;
 	}
 }
